Moved InputDateTime to its own file
